requirements.txt
streamlit
google-generativeai
python-dotenv
pypdf
cryptography
llama-index
llama-index-core
llama-index-llms-gemini
llama-index-embeddings-gemini
streamlit-pdf-viewer
docx2txt
pandas